move /charts under /vis
